fixed merge conflicts and changed the buttons in compact mode to vote buttons for consistency
